requirements.txt
streamlit==1.26.0
requests==2.31.0
fpdf==1.7.2
joblib==1.3.2
Pillow==9.4.0
scikit-learn==1.3.0
pandas==1.5.3
numpy==1.23.5
